% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayData.R
\name{displayData}
\alias{displayData}
\title{Basic Point Data Visualization}
\usage{
displayData(dataset, field)
}
\arguments{
\item{dataset}{Path to CSV file used as input}

\item{field}{Optional field that will enable informational popups upon hover (note that if no field is desired as input, an empty set of quotations must be specified (e.g., " "))}
}
\value{
A basic map showing the points based on the coordinates from the CSV file.
}
\description{
This function allows you to visualize an input point data file (as a CSV) on top of a basemap. You may additionally select a field to use for popups on the points, so the map becomes interactive. NOTE: the fields for latitude and longitude (to plot the points) must be labeled as "lat" and "lon" in your input CSV file (without the quotation marks)
}
\examples{
displayData(dataset = "../data/food.csv", field = STORE_NAME)
}
\keyword{data}
\keyword{point}
\keyword{visualization}
