% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTrend.R
\name{heatmapTrend}
\alias{heatmapTrend}
\title{Heatmap for Basic Trend Visualization}
\usage{
heatmapTrend(dataset, location)
}
\arguments{
\item{dataset}{Path to CSV file used as input}

\item{location}{Location, in quotations, of the geographical area in question}
}
\value{
A heatmap showing the trend in point data distribution
}
\description{
This function allows you to visualize an input point data file (as a CSV) on top of a basemap, in the form of a basic heatmap
}
\examples{
heatmapTrend(dataset = "../data/food.csv", field = STORE_NAME)
}
\keyword{data}
\keyword{heatmap,}
\keyword{visualization}
