% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showClusters.R
\name{showClusters}
\alias{showClusters}
\title{Clustering of Point Data}
\usage{
showClusters(dataset, field)
}
\arguments{
\item{dataset}{Path to CSV file used as input}

\item{field}{Optional field that will enable informational popups upon hover (note that if no field is desired as input, an empty set of quotations must be specified (e.g., " "))}
}
\value{
A map of the points clustered based on proximity
}
\description{
This function allows you to visualize an input point data file (as a CSV) on top of a basemap, as point clusters. As one zooms in, specific data points are revealed with their particular locations. You may additionally select a field to use for popups on the points, so the map becomes interactive. NOTE: the fields for latitude and longitude (coordinates to plot the points) must be labeled as "lat" and "lon" in your input CSV file (without the quotation marks)
}
\examples{
showClusters(dataset = "../data/food.csv", field = STORE_NAME)
}
\keyword{clustering,}
\keyword{data}
\keyword{point}
\keyword{visualization}
